Report water/target temperature as precise degC so HA's degF display is exact
The display->degC conversion rounded to whole degC, so a 103degF panel reading became 39degC,
and HA (degF) converted that back to 102degF -- 1degF lost in the round-trip. getCurrentTemperature/
getTargetTemperature/convertDisplayToCelsius now return a float and keep the precise degC
(39.44), so HA's degC->degF display lands back on the panel's exact degF. Unknown is NAN.

+++ components/intexsbh20/SBH20IO.h
@@ -74,8 +74,8 @@ public:
 public:
   bool isOnline() const;
 
-  int getCurrentTemperature() const;
-  int getTargetTemperature() const;
+  float getCurrentTemperature() const;
+  float getTargetTemperature() const;
   void forceReadTargetTemperature();
 
   uint8_t isBubbleOn() const;
@@ -211,7 +211,7 @@ private:
   static volatile uint32_t dbgIsrCalls, dbgLatchCalls;
 
 private:
-  uint16_t convertDisplayToCelsius(uint16_t value) const;
+  float convertDisplayToCelsius(uint16_t value) const;
   void processFrames();
   void tickControls(); // non-blocking control sequencer, called from loop()
